Generate complete contract
